AcctIdx: move background() to bucket holder
